requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
youtube-transcript-api>=0.6.0
requests>=2.31.0